requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
motor==3.3.2
pymongo==4.6.1
langchain==0.3.2
langchain-google-genai==2.0.0
langchain-mongodb==0.2.0
python-dotenv==1.0.1
pypdf==4.0.2 